pyproject.toml
[tool.poetry]
name = "ProjectFASTAPI"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.0"
pre-commit = "^3.3.3"
alembic = "^1.11.3"
psycopg2-binary = "^2.9.7"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pydantic-settings = "^2.0.3"
pandas = "^2.1.0"
numpy = "^1.25.2"
openpyxl = "^3.1.2"
xlsxwriter = "^3.1.2"
sqlalchemy-utils = "^0.41.1"
python-multipart = "^0.0.6"
boto3 = "^1.28.53"
requests = "^2.31.0"
pyjwt = "^2.8.0"
pydantic = {extras = ["email"], version = "^2.4.2"}
python-jose = "^3.3.0"
jinja2 = "^3.1.2"
typer = {extras = ["all"], version = "^0.9.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"